+++ cligen/argcvt.nim
@@ -30,42 +30,7 @@ proc nimEscape*(s: string, quote='"'): string =
 
 proc unescape*(s: string): string =
   ## Handles 1-byte octal, \[abtnvfre], and \xDD hex escapes
-  let hexdigits = { '0', '1', '2', '3', '4', '5', '6', '7', '8', '9',
-                    'a', 'b', 'c', 'd', 'e', 'f'}
-  proc toHexDig(c: char): int =
-    result = if c <= '9': ord(c) - ord('0') else: 10 + ord(c) - ord('a')
-  result = newStringOfCap(s.len)
-  var i = 0
-  while i < s.len:
-    case s[i]
-    of '\\':
-      if i + 1 >= s.len:
-        raise newException(ValueError, "incomplete escaped char")
-      case s[i+1]
-      of '0' .. '7': result.add(chr(ord(s[i+1]) - ord('0'))); inc(i, 2)
-      of 'a': result.add(chr(0x07)); inc(i, 2)       #bell
-      of 'b': result.add(chr(0x08)); inc(i, 2)       #backspace
-      of 't': result.add(chr(0x09)); inc(i, 2)       #horizontal tab
-      of 'n': result.add(chr(0x0A)); inc(i, 2)       #new line
-      of 'v': result.add(chr(0x0B)); inc(i, 2)       #vertical tab
-      of 'f': result.add(chr(0x0C)); inc(i, 2)       #form feed
-      of 'r': result.add(chr(0x0D)); inc(i, 2)       #carriage ret
-      of '\\': result.add(chr(0x5C)); inc(i, 2)      #backslash
-      of 'e': result.add(chr(0x1B)); inc(i, 2)       #escape
-      else:
-        if i + 3 >= s.len:
-          raise newException(ValueError, "incomplete 2-nibble hex escape")
-        if s[i+1].toLowerAscii != 'x':
-          raise newException(ValueError, "hex escape not of form \\xDD")
-        let dhi = toLowerAscii(s[i+2])
-        let dlo = toLowerAscii(s[i+3])
-        if dhi notin hexdigits or dlo notin hexdigits:
-          raise newException(ValueError, "non-hex escape: " & s[i..i+3])
-        result.add(char(toHexDig(dhi)*16 + toHexDig(dlo)))
-        inc(i, 4)
-    else:
-      result.add(s[i])
-      inc(i, 1)
+  for c in textUt.unescaped(s): result.add c
 
 type ElementError* = object of ValueError
 type ArgcvtParams* = object ## \